+++ customers/customers/db/base.py
@@ -1,4 +1,4 @@
-# Import all the models, so that Base has them before being
-# imported by Alembic
+# Import all the models, so that models.py won't crash on queries
 from customers.db.base_class import Base  # noqa
-from customers.models.customer import Address, Customer  # noqa
+from customers.models.customer import Customer  # noqa
+from customers.models.address import Address  # noqa