+++ customers/customers/models/__init__.py
@@ -1,0 +1,3 @@
+# Import all the models, so that models.py won't crash on queries
+from customers.models.customer import Customer  # noqa
+from customers.models.address import Address  # noqa